# docs/openapi-c-v1.yaml
openapi: 3.0.3
info:
  title: Baoyunlian C API
  version: 1.0.0
  description: Frozen C-side API contract for frontend-backend integration.
servers:
  - url: http://127.0.0.1:4000
security:
  - bearerAuth: []
...
paths:
...
  /api/mall/redeem:
    post:
      tags: [Mall]
      summary: Redeem item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                itemId: { type: integer }
              required: [itemId]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  token: { type: string }
                  balance: { type: integer }
                required: [ok, token, balance]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/redemptions/{id}/writeoff:
    post:
      tags: [Redemptions]
      summary: Writeoff redemption
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                required: [ok]
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410':
          description: Gone
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorBody' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
...
  schemas:
    ErrorBody:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
      required: [code, message]
...